Rename util.mjs to util.js
