+++ src/instruments/src/EFB/index.jsx
@@ -21,7 +21,7 @@ import { useState, useEffect } from 'react';
 import Efb from './Efb.tsx';
 import {
     renderTarget, getSimVar, setSimVar,
-} from '../util.mjs';
+} from '../util.js';
 import logo from './Assets/fbw-logo.svg';
 
 import './Assets/Efb.scss';
